[DEV] | devide tracemodel_library
