+++ main/tracemodel_library.cpp
@@ -1,28 +1,3 @@
-#ifndef _TRACEMODEL_H_
-#define _TRACEMODEL_H_
-#include <SoftwareSerial.h>
-#include <AFMotor.h>
-
-class TraceModel{
-private:
-  int speed_fro;
-  int speed_cur;
-  int line_r = A0;
-  int line_l = A5;
-  AF_DCMotor* pm1;
-  AF_DCMotor* pm2;
-
-  int stat[2];
-  int speedArr[2];
-
-  int isChange(int type, int speed);
-  void Move(char cmd);
-  void Speed(int type, int speed);
-public:
-  TraceModel(int sp_f, int sp_c, int motor1_pin, int motor2_pin);
-  void Go();
-};
-
 TraceModel :: TraceModel(int sp_f, int sp_c, int motor1_pin, int motor2_pin){
   pm1 = new AF_DCMotor(motor1_pin);
   pm2 = new AF_DCMotor(motor2_pin);
@@ -144,5 +119,3 @@ void TraceModel :: Speed(int type, int speed){
   }
   //delay(100);
 }
-
-#endif
